libext2fs: avoid too much memory allocation in case fs_num_threads
e2fsck init memory according to filesystem inodes/dir numbers
recorded in the superblock, this should be aware of filesystem
number of threads, otherwise, oom can happen.

So in case of fs->fs_num_threads, this patch controls the amount of
memory consumed for running multiple threads in e2fsck.

Signed-off-by: Wang Shilong <wshilong@ddn.com>
Signed-off-by: Ritesh Harjani (IBM) <ritesh.list@gmail.com>